+++ src/main/java/firok/tiths/intergration/conarm/traits/TraitArmorChemicalInstable.java
@@ -1,9 +1,7 @@
 package firok.tiths.intergration.conarm.traits;
 
-import firok.tiths.common.Items;
 import firok.tiths.intergration.conarm.IAbstractArmorTrait;
-import firok.tiths.traits.TraitCarbonizing;
-import firok.tiths.util.Actions;
+import firok.tiths.traits.TraitChemicalInstable;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.DamageSource;
@@ -12,16 +10,16 @@ import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import static firok.tiths.util.Predicates.canTrigger;
 
 /**
- * 碳化 - 护甲
+ * 化学不稳定 - 护甲
  */
-public class TraitArmorCarbonizing extends TraitCarbonizing implements IAbstractArmorTrait
+public class TraitArmorChemicalInstable extends TraitChemicalInstable implements IAbstractArmorTrait
 {
 	@Override
 	public float onHurt(ItemStack armor, EntityPlayer player, DamageSource source, float damage, float newDamage, LivingHurtEvent evt)
 	{
-		if(!player.world.isRemote && canTrigger(player.world,0.08f))
+		if(canTrigger(player.world,0.12f))
 		{
-			Actions.CauseSpawnItem(player,new ItemStack(Items.cinder));
+			boom(player,player.posX,player.posY,player.posZ);
 		}
 		return newDamage;
 	}