+++ src/main/java/firok/tiths/intergration/conarm/traits/TraitArmorDecoying.java
@@ -1,8 +1,7 @@
 package firok.tiths.intergration.conarm.traits;
 
-import firok.tiths.common.Items;
 import firok.tiths.intergration.conarm.IAbstractArmorTrait;
-import firok.tiths.traits.TraitCarbonizing;
+import firok.tiths.traits.TraitDecoying;
 import firok.tiths.util.Actions;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
@@ -12,16 +11,16 @@ import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import static firok.tiths.util.Predicates.canTrigger;
 
 /**
- * 碳化 - 护甲
+ * 诱食 - 护甲
  */
-public class TraitArmorCarbonizing extends TraitCarbonizing implements IAbstractArmorTrait
+public class TraitArmorDecoying extends TraitDecoying implements IAbstractArmorTrait
 {
 	@Override
 	public float onHurt(ItemStack armor, EntityPlayer player, DamageSource source, float damage, float newDamage, LivingHurtEvent evt)
 	{
-		if(!player.world.isRemote && canTrigger(player.world,0.08f))
+		if(!player.world.isRemote && canTrigger(player.world,rate))
 		{
-			Actions.CauseSpawnItem(player,new ItemStack(Items.cinder));
+			Actions.CauseSpawningPassives(player);
 		}
 		return newDamage;
 	}