+++ src/main/java/firok/tiths/intergration/conarm/traits/TraitArmorDichroic.java
@@ -1,0 +1,12 @@
+package firok.tiths.intergration.conarm.traits;
+
+import firok.tiths.intergration.conarm.IAbstractArmorTrait;
+import firok.tiths.traits.TraitDichroic;
+
+/**
+ * 二色性 - 护甲
+ */
+public class TraitArmorDichroic extends TraitDichroic implements IAbstractArmorTrait
+{
+	// todo
+}