+++ src/main/java/firok/tiths/intergration/conarm/traits/TraitArmorGluttonic.java
@@ -1,0 +1,12 @@
+package firok.tiths.intergration.conarm.traits;
+
+import firok.tiths.intergration.conarm.IAbstractArmorTrait;
+import firok.tiths.traits.TraitGluttonic;
+
+/**
+ * 暴食 - 护甲
+ */
+public class TraitArmorGluttonic extends TraitGluttonic implements IAbstractArmorTrait
+{
+	// todo
+}