+++ src/main/java/firok/tiths/intergration/conarm/traits/TraitArmorGorgeous.java
@@ -1,28 +1,20 @@
 package firok.tiths.intergration.conarm.traits;
 
-import firok.tiths.common.Items;
 import firok.tiths.intergration.conarm.IAbstractArmorTrait;
-import firok.tiths.traits.TraitCarbonizing;
-import firok.tiths.util.Actions;
+import firok.tiths.traits.TraitGorgeous;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.DamageSource;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
 
-import static firok.tiths.util.Predicates.canTrigger;
-
 /**
- * 碳化 - 护甲
+ * 斑斓 - 护甲
  */
-public class TraitArmorCarbonizing extends TraitCarbonizing implements IAbstractArmorTrait
+public class TraitArmorGorgeous extends TraitGorgeous implements IAbstractArmorTrait
 {
 	@Override
 	public float onHurt(ItemStack armor, EntityPlayer player, DamageSource source, float damage, float newDamage, LivingHurtEvent evt)
 	{
-		if(!player.world.isRemote && canTrigger(player.world,0.08f))
-		{
-			Actions.CauseSpawnItem(player,new ItemStack(Items.cinder));
-		}
-		return newDamage;
+		return newDamage / factor(player);
 	}
 }