+++ changelog.md
@@ -2,6 +2,15 @@
 
 > All Time Stamps: GMT+8
 
+# 0.2.58.0 _2020-03-23 13:30_
+
+* 为属性增加护甲效果
+  * 化学不稳定 chemical instable
+  * 喀嚓 creaky
+  * 诱食 decoying
+  * 极寒 extreme freezing
+  * 斑斓 gorgeous
+
 # 0.2.57.0 _2020-03-20 17:30_
 
 * 增加物品 added items